eventhub_backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
boto3==1.35.87
botocore==1.35.99
celery==5.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.0.12
django-anymail==12.0
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.3
django-ratelimit==4.1.0
django-storages==1.14.4
django-timezone-field==7.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
factory_boy==3.3.1
Faker==40.5.1
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
packaging==26.0
pillow==11.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.2.5
psycopg-binary==3.2.5
PyJWT==2.11.0
pypng==0.20220715.0
pytest==8.3.4
pytest-django==4.10.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
qrcode==7.4.2
redis==5.0.8
referencing==0.37.0
requests==2.32.3
rpds-py==0.30.0
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.5
stripe==11.5.0
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.7.0
